ml/requirements.txt
torch>=2.4.0
torchvision>=0.19.0
timm>=1.0.0
datasets>=3.0.0
huggingface-hub>=0.24.0
Pillow>=11.0.0
numpy>=2.2.0
scikit-learn>=1.7.0
scipy>=1.15.0
umap-learn>=0.5.7
matplotlib>=3.10.0
tqdm>=4.67.0

